refactor: move addProduct api to product package
