+++ greengift/src/main/java/com/apiserver/greengift/festival/FestivalRestController.java
@@ -26,14 +26,6 @@ public class FestivalRestController {
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
-    @PostMapping("/product/{festivalId}")
-    public ResponseEntity<?> addProduct(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                        @RequestBody @Valid FestivalRequest.AddProduct requestDTO,
-                                        @PathVariable @Min(1) Long festivalId) {
-        festivalService.addProduct(userDetails.getUser(), requestDTO, festivalId);
-        return ResponseEntity.ok().body(ApiUtils.success(null));
-    }
-
     @PostMapping("/join")
     public ResponseEntity<?> joinFestival(@AuthenticationPrincipal CustomUserDetails userDetails,
                                           @RequestBody @Valid FestivalRequest.JoinFestival requestDTO) {
@@ -43,14 +35,14 @@ public class FestivalRestController {
 
     @GetMapping("/random/{festivalId}")
     public ResponseEntity<?> findRandomResult(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                        @PathVariable @Min(1) Long festivalId) {
+                                              @PathVariable @Min(1) Long festivalId) {
         festivalService.findRandomResult(userDetails.getUser(), festivalId);
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
     @GetMapping("/result/{festivalId}")
     public ResponseEntity<?> findFestivalResult(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                               @PathVariable @Min(1) Long festivalId) {
+                                                @PathVariable @Min(1) Long festivalId) {
         FestivalResponse.FindFestivalResult response = festivalReadService.findFestivalResult(userDetails.getUser(), festivalId);
         return ResponseEntity.ok().body(ApiUtils.success(response));
     }
